postprocessor-python-objects-sender/requirements.txt
pyinstaller
msgpack
scipy
numpy
opencv-python
requests
torch
open_clip_torch
qdrant-client
Pillow